+++ example/lib/main.dart
@@ -1,17 +1,21 @@
 import 'dart:async';
 
 import 'package:example/detail.dart';
-import 'package:example/utils/sp_client.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_mana/flutter_mana.dart';
 import 'package:flutter_mana_kits/flutter_mana_kits.dart';
 
+import 'utils/animated_ball.dart';
+import 'utils/custom_button.dart';
+import 'utils/custom_text_field.dart';
 import 'utils/dio_client.dart';
 import 'utils/log_generator.dart';
+import 'utils/sp_client.dart';
 
 void main() async {
   ManaPluginManager.instance
+    ..register(ManaVisualHelper())
     ..register(ManaGrid())
     ..register(ManaLicense())
     ..register(ManaPackageInfo())
@@ -80,6 +84,8 @@ class _HomePageState extends State<HomePage> with TickerProviderStateMixin {
 
   @override
   Widget build(BuildContext context) {
+    print('build');
+
     return Scaffold(
       backgroundColor: Colors.white,
       appBar: AppBar(backgroundColor: Theme.of(context).colorScheme.inversePrimary, title: Text(widget.title)),
@@ -105,58 +111,13 @@ class _HomePageState extends State<HomePage> with TickerProviderStateMixin {
                 backgroundColor: Colors.deepPurple,
                 onPressed: addSharedPreferences,
               ),
+              CommonTextField(initialValue: '文本测试'),
+              AnimatedBall(),
+              SizedBox(width: 300, child: Image.asset('assets/test.jpeg', fit: BoxFit.cover)),
             ],
           ),
         ),
       ),
-    );
-  }
-}
-
-class CustomButton extends StatefulWidget {
-  final String text;
-  final Color? backgroundColor;
-  final Color? foregroundColor;
-  final FutureOr<void> Function()? onPressed;
-
-  const CustomButton({
-    super.key,
-    required this.text,
-    this.backgroundColor,
-    this.foregroundColor = Colors.white,
-    this.onPressed,
-  });
-
-  @override
-  State<CustomButton> createState() => _CustomButtonState();
-}
-
-class _CustomButtonState extends State<CustomButton> {
-  bool _isLoading = false;
-
-  Future<void> _handlePress() async {
-    if (_isLoading || widget.onPressed == null) return;
-
-    setState(() => _isLoading = true);
-
-    try {
-      final result = widget.onPressed!();
-      if (result is Future) {
-        await result;
-      }
-    } finally {
-      if (mounted) {
-        setState(() => _isLoading = false);
-      }
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return ElevatedButton(
-      onPressed: _isLoading ? null : _handlePress,
-      style: ElevatedButton.styleFrom(backgroundColor: widget.backgroundColor, foregroundColor: widget.foregroundColor),
-      child: Text(widget.text, style: const TextStyle(fontSize: 14, fontWeight: FontWeight.bold)),
     );
   }
 }