+++ flutter_mana_kits/lib/src/plugins/mana_align_ruler/widgets/selected_rect_painter.dart
@@ -19,8 +19,8 @@ class SelectedRectPainter extends CustomPainter {
 
     final Matrix4 transform = selectedObject!.getTransformTo(null);
 
-    final Rect globalRect = selectedObject!.paintBounds.shift(
-        Offset(transform.getTranslation().x, transform.getTranslation().y));
+    final Rect globalRect =
+        selectedObject!.paintBounds.shift(Offset(transform.getTranslation().x, transform.getTranslation().y));
 
     final Paint fillPaint = Paint()
       ..color = fillColor.withAlpha(50)