+++ flutter_mana_kits/lib/src/plugins/mana_color_sucker/widgets/color_sucker_barrier.dart
@@ -95,17 +95,19 @@ class _ColorSuckerBarrierState extends State<ColorSuckerBarrier> {
       onPanEnd: (details) => _onUpdate(details.globalPosition, true),
       child: Stack(
         children: [
+          Positioned.fill(
+              child: Container(
+            color: Colors.transparent,
+          )),
           Positioned(
             left: dragPosition.dx - widget.magnifierRadius,
             top: dragPosition.dy - widget.magnifierRadius,
             child: RawMagnifier(
-              size:
-                  Size(widget.magnifierRadius * 2, widget.magnifierRadius * 2),
+              size: Size(widget.magnifierRadius * 2, widget.magnifierRadius * 2),
               magnificationScale: widget.magnification,
               decoration: MagnifierDecoration(
                 shape: CircleBorder(
-                  side: BorderSide(
-                      color: widget.borderColor, width: widget.borderWidth),
+                  side: BorderSide(color: widget.borderColor, width: widget.borderWidth),
                 ),
               ),
               focalPointOffset: Offset.zero,