+++ flutter_mana_kits/lib/src/plugins/mana_log_viewer/widgets/log_viewer_setting.dart
@@ -5,8 +5,7 @@ class LogViewerSetting extends StatelessWidget with I18nMixin {
   final bool verboseLogs;
   final ValueChanged<bool?>? onChanged;
 
-  const LogViewerSetting(
-      {super.key, required this.verboseLogs, this.onChanged});
+  const LogViewerSetting({super.key, required this.verboseLogs, this.onChanged});
 
   @override
   Widget build(BuildContext context) {