+++ flutter_mana_kits/lib/src/plugins/mana_memory_info/widgets/memory_detail.dart
@@ -22,8 +22,7 @@ class MemoryDetail extends StatelessWidget with I18nMixin {
   /// 从 MemoryService 获取类的详细信息。
   Future<ClsModel?> _fetchClassDetails() async {
     final completer = Completer<ClsModel?>();
-    service.getClassDetailInfo(
-        detail.classRef?.id ?? '', (info) => completer.complete(info));
+    service.getClassDetailInfo(detail.classRef?.id ?? '', (info) => completer.complete(info));
     return completer.future;
   }
 
@@ -48,14 +47,11 @@ class MemoryDetail extends StatelessWidget with I18nMixin {
           future: _fetchClassDetails(),
           builder: (context, snapshot) {
             if (!snapshot.hasData || snapshot.data == null) {
-              return Center(
-                  child: Text(t('memory_info.class_no_detail'),
-                      style: TextStyle(fontSize: 20)));
+              return Center(child: Text(t('memory_info.class_no_detail'), style: TextStyle(fontSize: 20)));
             }
 
             final info = snapshot.data!;
-            final propertiesText =
-                info.properties.map((e) => e.propertyStr).join('\n');
+            final propertiesText = info.properties.map((e) => e.propertyStr).join('\n');
             final functionsText = info.functions.join('\n');
 
             return SingleChildScrollView(
@@ -69,8 +65,7 @@ class MemoryDetail extends StatelessWidget with I18nMixin {
                       padding: EdgeInsets.only(bottom: 10),
                       child: Text(
                         t('memory_info.class_name'),
-                        style: TextStyle(
-                            fontSize: 20, fontWeight: FontWeight.bold),
+                        style: TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
                       ),
                     ),
                     Text(
@@ -83,8 +78,7 @@ class MemoryDetail extends StatelessWidget with I18nMixin {
                       padding: EdgeInsets.only(bottom: 10),
                       child: Text(
                         t('memory_info.location'),
-                        style: TextStyle(
-                            fontSize: 20, fontWeight: FontWeight.bold),
+                        style: TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
                       ),
                     ),
                     Text(
@@ -97,8 +91,7 @@ class MemoryDetail extends StatelessWidget with I18nMixin {
                       padding: EdgeInsets.only(bottom: 10),
                       child: Text(
                         t('memory_info.property'),
-                        style: TextStyle(
-                            fontSize: 20, fontWeight: FontWeight.bold),
+                        style: TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
                       ),
                     ),
                     Text(
@@ -111,8 +104,7 @@ class MemoryDetail extends StatelessWidget with I18nMixin {
                       padding: EdgeInsets.only(bottom: 10),
                       child: Text(
                         t('memory_info.function'),
-                        style: TextStyle(
-                            fontSize: 20, fontWeight: FontWeight.bold),
+                        style: TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
                       ),
                     ),
                     Text(