+++ flutter_mana_kits/lib/src/plugins/mana_memory_info/widgets/memory_info.dart
@@ -17,6 +17,7 @@ class MemoryInfo extends StatelessWidget {
       position: PositionType.bottom,
       drag: false,
       showBarrier: false,
+      maintainContent: false,
       content: MemoryInfoContent(),
     );
   }