+++ flutter_mana_kits/lib/src/plugins/mana_memory_info/widgets/memory_info_content.dart
@@ -27,16 +27,11 @@ class _MemoryInfoContentState extends State<MemoryInfoContent> with I18nMixin {
     var items = _memoryService.classHeapStatsList;
 
     if (_hidePrivateClasses) {
-      items = items
-          .where((v) => !(v.classRef?.name?.startsWith('_') ?? false))
-          .toList();
+      items = items.where((v) => !(v.classRef?.name?.startsWith('_') ?? false)).toList();
     }
 
     if (_onlyCurrentPackage && _pkg.isNotEmpty) {
-      items = items
-          .where((v) =>
-              (v.classRef?.location?.script?.uri?.contains(_pkg) ?? false))
-          .toList();
+      items = items.where((v) => (v.classRef?.location?.script?.uri?.contains(_pkg) ?? false)).toList();
     }
 
     return items;
@@ -98,11 +93,9 @@ class _MemoryInfoContentState extends State<MemoryInfoContent> with I18nMixin {
     });
   }
 
-  Widget _buildTableHeaderCol(
-      {int flex = 1, required String field, required String label}) {
-    final icon = _currentSortField == field
-        ? (_isAscending ? KitIcons.sort_asc : KitIcons.sort_desc)
-        : KitIcons.sort_auto;
+  Widget _buildTableHeaderCol({int flex = 1, required String field, required String label}) {
+    final icon =
+        _currentSortField == field ? (_isAscending ? KitIcons.sort_asc : KitIcons.sort_desc) : KitIcons.sort_auto;
 
     return Expanded(
       flex: flex,
@@ -110,9 +103,7 @@ class _MemoryInfoContentState extends State<MemoryInfoContent> with I18nMixin {
         onTap: () => _onSortChanged(field),
         child: Row(
           children: [
-            Text(t(label),
-                style:
-                    const TextStyle(fontSize: 16, fontWeight: FontWeight.bold)),
+            Text(t(label), style: const TextStyle(fontSize: 16, fontWeight: FontWeight.bold)),
             Icon(icon, size: 16),
           ],
         ),
@@ -124,11 +115,9 @@ class _MemoryInfoContentState extends State<MemoryInfoContent> with I18nMixin {
     return Row(
       crossAxisAlignment: CrossAxisAlignment.start,
       children: [
-        Text(label,
-            style: const TextStyle(fontSize: 14, fontWeight: FontWeight.bold)),
+        Text(label, style: const TextStyle(fontSize: 14, fontWeight: FontWeight.bold)),
         const SizedBox(width: 8),
-        Expanded(
-            child: SelectableText(value, style: const TextStyle(fontSize: 14))),
+        Expanded(child: SelectableText(value, style: const TextStyle(fontSize: 14))),
       ],
     );
   }
@@ -171,8 +160,7 @@ class _MemoryInfoContentState extends State<MemoryInfoContent> with I18nMixin {
             expandedHeight: 420.0,
             pinned: true,
             flexibleSpace: FlexibleSpaceBar(
-              titlePadding:
-                  const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
+              titlePadding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
               background: Container(
                 color: Colors.white,
                 padding: const EdgeInsets.all(16),
@@ -181,34 +169,26 @@ class _MemoryInfoContentState extends State<MemoryInfoContent> with I18nMixin {
                   crossAxisAlignment: CrossAxisAlignment.start,
                   spacing: 2,
                   children: [
-                    Text(t('memory_info.vm_info'),
-                        style: const TextStyle(
-                            fontSize: 18, fontWeight: FontWeight.bold)),
+                    Text(t('memory_info.vm_info'), style: const TextStyle(fontSize: 18, fontWeight: FontWeight.bold)),
                     const SizedBox(height: 8),
                     _buildTableRow('PID:', '${vmInfo?.pid ?? 'N/A'}'),
                     _buildTableRow('CPU:', vmInfo?.hostCPU ?? 'N/A'),
-                    _buildTableRow(
-                        t('memory_info.version'), vmInfo?.version ?? 'N/A'),
+                    _buildTableRow(t('memory_info.version'), vmInfo?.version ?? 'N/A'),
                     const SizedBox(height: 16),
                     Text(t('memory_info.memory_info'),
-                        style: const TextStyle(
-                            fontSize: 18, fontWeight: FontWeight.bold)),
+                        style: const TextStyle(fontSize: 18, fontWeight: FontWeight.bold)),
                     const SizedBox(height: 8),
                     _buildTableRow(
                         'Heap Usage:',
                         '${memoryInfo?.heapUsageFormatted ?? 'N/A'} / '
                             '${memoryInfo?.heapCapacityFormatted ?? 'N/A'}'),
-                    _buildTableRow('External Usage:',
-                        memoryInfo?.externalUsageFormatted ?? 'N/A'),
+                    _buildTableRow('External Usage:', memoryInfo?.externalUsageFormatted ?? 'N/A'),
                     const SizedBox(height: 16),
-                    _buildCheckboxRow(
-                        _hidePrivateClasses, 'memory_info.hide_private_classes',
-                        () {
+                    _buildCheckboxRow(_hidePrivateClasses, 'memory_info.hide_private_classes', () {
                       _hidePrivateClasses = !_hidePrivateClasses;
                     }),
                     const SizedBox(height: 4),
-                    _buildCheckboxRow(
-                        _onlyCurrentPackage, 'memory_info.current_app', () {
+                    _buildCheckboxRow(_onlyCurrentPackage, 'memory_info.current_app', () {
                       _onlyCurrentPackage = !_onlyCurrentPackage;
                     }),
                     const SizedBox(height: 4),
@@ -221,18 +201,12 @@ class _MemoryInfoContentState extends State<MemoryInfoContent> with I18nMixin {
               preferredSize: const Size.fromHeight(39),
               child: Container(
                 color: Colors.grey.shade200,
-                padding:
-                    const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
+                padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
                 child: Row(
                   children: [
-                    _buildTableHeaderCol(
-                        field: 'accumulatedSize', label: 'memory_info.size'),
-                    _buildTableHeaderCol(
-                        field: 'instancesCurrent', label: 'memory_info.number'),
-                    _buildTableHeaderCol(
-                        flex: 3,
-                        field: 'className',
-                        label: 'memory_info.class_name'),
+                    _buildTableHeaderCol(field: 'accumulatedSize', label: 'memory_info.size'),
+                    _buildTableHeaderCol(field: 'instancesCurrent', label: 'memory_info.number'),
+                    _buildTableHeaderCol(flex: 3, field: 'className', label: 'memory_info.class_name'),
                   ],
                 ),
               ),
@@ -247,22 +221,16 @@ class _MemoryInfoContentState extends State<MemoryInfoContent> with I18nMixin {
                 return InkWell(
                   onTap: () => Navigator.of(context).push(
                     MaterialPageRoute(
-                      builder: (_) =>
-                          MemoryDetail(detail: item, service: _memoryService),
+                      builder: (_) => MemoryDetail(detail: item, service: _memoryService),
                     ),
                   ),
                   child: Container(
                     color: isEven ? Colors.grey.shade100 : Colors.white,
-                    padding:
-                        const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
+                    padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
                     child: Row(
                       children: [
-                        Expanded(
-                            flex: 1,
-                            child: Text(item.accumulatedSizeFormatted)),
-                        Expanded(
-                            flex: 1,
-                            child: Text('${item.instancesCurrent ?? 'N/A'}')),
+                        Expanded(flex: 1, child: Text(item.accumulatedSizeFormatted)),
+                        Expanded(flex: 1, child: Text('${item.instancesCurrent ?? 'N/A'}')),
                         Expanded(
                           flex: 3,
                           child: Text(