+++ flutter_mana_kits/lib/src/plugins/mana_package_info/widgets/package_info.dart
@@ -12,6 +12,7 @@ class PackageInfo extends StatelessWidget {
   Widget build(BuildContext context) {
     return ManaFloatingWindow(
       name: name,
+      showBarrier: false,
       content: PackageInfoContent(),
     );
   }