+++ flutter_mana_kits/lib/src/plugins/mana_shared_preferences_viewer/index.dart
@@ -8,8 +8,7 @@ class ManaSharedPreferencesViewer implements ManaPluggable {
   const ManaSharedPreferencesViewer();
 
   @override
-  Widget? buildWidget(BuildContext? context) =>
-      SharedPreferencesViewer(name: name);
+  Widget? buildWidget(BuildContext? context) => SharedPreferencesViewer(name: name);
 
   @override
   String getLocalizedDisplayName(Locale locale) {